+++ spec/assessments/keywordStopWordsSpec.js
