+++ spec/assessments/textCompetingLinksSpec.js
@@ -1,0 +1,10 @@
+var linkStatisticAssessment = require( "../../js/assessments/textCompetingLinksAssessment.js" );
+var Paper = require( "../../js/values/Paper.js" );
+
+var factory = require( "../helpers/factory.js" );
+var i18n = factory.buildJed();
+
+describe( "An assessor running the linkStatistics", function(){
+	
+
+} );
