+++ spec/assessments/textImagesSpec.js
