+++ spec/assessments/textLengthSpec.js
